[proposal] Structured Async for Mojo
Proposes to add structured async to Mojo, following in the the Rust
tradition of async since Mojo has the ability to fix many of the issues
with Rust's async, some of which are ecosystem inflicted and some of
which are inflicted by old decisions made about the language (such as
that any value may be leaked). I think that this interface offers a lot
more flexibility than the current one does for high performance code,
while providing a better path for gradual evolution. It does have some
language-level dependences, namely unions, and requires the introduction
of `Send` and `Sync` traits, which are used to control data movement
between threads.

Signed-off-by: Owen Hilyard <hilyard.owen@gmail.com>