[proposal] Provided Effect Handlers
This proposal contains an alternative to an effect system which I think
is more suitable for abstracting `async`, `raises`, and similar function
colors in a systems language where the context may not always be suited
to running arbitrary code. This is done by inverting the normal effect
system, and having libraries provide the handlers based on information
passed to them by the caller. Aside from the ability to manipulate
function signatures with parameters (`async`, `raises`, and the return
type), this can be implemented in Mojo today, but has substantial
ergonomics issues. As such, I propose a place to put implicit parameters
that not all functions will care about but should be propagated, such as
division by zero handling, FP error handling, OOM, error handling
behavior and `async`. With this, only functions which actually perform
IO need to deal with async beyond awaiting functions they call (which
does nothing for sync functions).

Signed-off-by: Owen Hilyard <hilyard.owen@gmail.com>